Double

+++ Ejercicio 3/complex.h
@@ -4,12 +4,12 @@
 
 class Complex final : public Number {
     private:
-        const float realValue;
-        const float imaginaryValue;
+        const double realValue;
+        const double imaginaryValue;
         
     public:
         Complex() : realValue(0), imaginaryValue(0) {};
-        Complex (const float _realValue, const float _imaginaryValue) : realValue(_realValue), imaginaryValue(_imaginaryValue) {};
+        Complex (const double _realValue, const double _imaginaryValue) : realValue(_realValue), imaginaryValue(_imaginaryValue) {};
 
         unique_ptr<Number> operator+(const Number& number) const override;
         unique_ptr<Number> operator-(const Number& number) const override;
